requirements.txt
pandas
openpyxl
python-docx
aspose-words
